Brings the PIX surplus onto the wire and the single `PixParams` encoding that
carries it.

`SsaDimensions` is folded away here too, which needed a decision this branch had
already made differently: on `lukas/pix` the type was near-dead, but the
supervisor gave it a job — it is the `dims` field of `SessionPixSupervisor`, and
`target_useful_shares()` hangs off it. That method moves onto `PixParams`, whose
own construction excludes the surplus by definition: a surplus share is one that
arrives after its polynomial is already complete, so it advances nothing and was
never part of the target. The supervisor's arithmetic is unchanged.

`SessionPixAction::RequestSsa` carries the triple rather than a `polys` /
`threshold` pair, so the dimensions reach `send_ssa_request` in one piece and the
failure path no longer has to invent a zero for them.

Conflict resolution of note: this branch deleted the manager tests for SSA index
allocation, batching and pipelining when the supervisor took those over, and
`lukas/pix` had edited them in place. The deletions win — the invariants live in
the supervisor's own suite now. `manager.rs:6178` was the one hazard the merge
could not flag: a hand-rolled `additional_data` literal that still compiled but
decoded to `shares = 0` under the new layout, failing a supervisor test for a
reason that had nothing to do with it. It now goes through the same helper as
every other test.

+++ protocols/hopr/src/codec/mod.rs
@@ -379,7 +379,7 @@ mod tests {
         // Need to generate multiple messages to recover the SSA
         let threshold = f.sender.ssa_gen.config().threshold as usize;
         let num_polys = f.sender.ssa_gen.config().polynomials_per_ssa as usize;
-        let surplus = f.sender.ssa_gen.config().surplus_shares;
+        let surplus = f.sender.ssa_gen.config().surplus_shares as usize;
         // Shares are emitted round-robin across the emission window, so every polynomial reaches
         // `threshold` inside the same round and the SSA completes on the last share of round
         // `threshold` — exactly the `polys × threshold` packets the agreed quota is priced on.
@@ -499,7 +499,7 @@ mod tests {
 
         let threshold = f.sender.ssa_gen.config().threshold as usize;
         let num_polys = f.sender.ssa_gen.config().polynomials_per_ssa as usize;
-        let surplus = f.sender.ssa_gen.config().surplus_shares;
+        let surplus = f.sender.ssa_gen.config().surplus_shares as usize;
         let shares_per_poly = threshold + surplus;
         let msgs_per_commitment = num_polys * shares_per_poly;
